Adds subregion support (shufting) to DctTiler.
The previous commit attempted this via support built into
GrpcNddiDisplay. That was a disaster, because A) only coordinates
in the CoefficientPlanes should be shifted, and B) sometimes the
coefficients themselves need to to be shifted (e.g. tx, ty) which
requires application domain knowledge.

Now, the smarts is in the application (PixelBridge), specifically
the tilers (DctTiler).